Update VirtualSwitchParent.groovy
The code from jwsf requires the irrigation controller to have a specific name for it to be recognized and selected to work with the VirtualSwitchParent app.  

I modified line 19 to allow the user to see all switches registered with their hub, including the Smarter Sprinkler project Irrigation Controller which uses the switch capability.